Run postinst script template generation when dirs
A regression was introduced for a previously untested feature in v4.3.0.
PR#203 no longer runs template generation if there are no commands,
which works very well for preinst, prerm and postrm. The postinst
template also has behavior for 'dirs' implemented, which would not run
in v4.3.0 unless the user had also defined postInst commands.

This applies to debians only.