Avoid scriptlet generation; special-case %pretrans.
Currently, scriptlets are generated for all stages regardless of whether
they do anything. This is best avoided.

Additionally, the %pretrans scriptlet is being treated the same as any
other type of scriptlet. This is a bad idea, as there are conditions
where a shell might not be present to run the %pretrans scriptlet.
Moreover, it's recommended in the Fedora Packaging Guidelines[1] that if
a %pretrans scriptlet is required, it MUST be written in Lua. This means
it makes no sense to include the preamble added to the scriptlet by
scriptWithUtils.

[1] https://fedoraproject.org/wiki/Packaging:Scriptlets#The_.25pretrans_Scriptlet