Make Deb and Rpm tasks cacheable
Both tasks already support up-to-date check so only the annotation was
missing to make them cacheable.

By default, all the built-in archiving tasks are not cacheable as they
are usually quick so it's not worth it to make them cacheable. The
difference with Rpm and Deb is the output produced, if task is executed
twice from a clean workspace, is different because a time stamp is
inserted in the deb or rpm.

Making those task cacheable is worth it to not invalidate the tasks that
depends on the deb or rpm, they will be retrieved from the gradle cache
if it is enabled.